NECB 2011 QAQC implementation (#350)
* reorganize necb 2011 and 2015 folder structure

* modified which necb 2015 qaqc tests runs

* fix incorrect final qaqc output

* move qaqc methods; minor change to qaqc result wording

* minor changes to qaqc class structure
