+++ lib/openstudio-standards/standards/necb/necb_2011/data/necb_2011_spacetype_info.csv
