+++ lib/openstudio-standards/standards/necb/necb_2011/data/space_types.json
