+++ lib/openstudio-standards/standards/necb/necb_2011/qaqc/qaqc_data/economizer.json
@@ -1,0 +1,18 @@
+{
+  "tables": [
+    {
+      "name": "economizer_compliance",
+      "data_type": "table",
+      "refs": [
+        "NECB2011-5.2.2.7"
+      ],
+      "table": [
+	    {
+		  "var": "economizer_control_type",
+		  "expected_value": "(capacity > 20000 or max_air_flow_rate_m3_per_s*1000 >1500) ? 'DifferentialEnthalpy' : 'NoEconomizer'",
+		  "note": "capacity can be either [dx single speed]'s `nominal total capacity` or [dx two speed]'s `highest cop`"
+		}
+	  ]
+	}
+  ]
+}
