+++ lib/openstudio-standards/standards/necb/necb_2011/qaqc/qaqc_data/exterior_fenestration.json
@@ -1,0 +1,159 @@
+{
+  "tables": [
+    {
+      "name": "exterior_fenestration_compliance",
+      "data_type": "table",
+      "refs": [
+        "NECB2011-Section 3.2.2.3"
+      ],
+      "table": [
+	    {
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.400,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_window_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 1.600,
+		  "climate_index": 5
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.400,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 1.600,
+		  "climate_index": 5
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.400,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_overhead_door_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 1.600,
+		  "climate_index": 5
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.400,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 2.200,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_skylight_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 1.600,
+		  "climate_index": 5
+		}
+		
+		
+	  ]
+	}
+  ]
+}
