+++ lib/openstudio-standards/standards/necb/necb_2011/qaqc/qaqc_data/exterior_opaque.json
@@ -1,0 +1,121 @@
+{
+  "tables": [
+    {
+      "name": "exterior_opaque_compliance",
+      "data_type": "table",
+      "refs": [
+        "NECB2011-Section 3.2.2.2"
+      ],
+      "table": [
+	    {
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.227,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.183,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.183,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.162,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.162,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_floor_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.142,
+		  "climate_index": 5
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.227,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.183,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.183,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.162,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.162,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_roof_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.142,
+		  "climate_index": 5
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.315,
+		  "climate_index": 0
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.278,
+		  "climate_index": 1
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.247,
+		  "climate_index": 2
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.210,
+		  "climate_index": 3
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.210,
+		  "climate_index": 4
+		},
+		{
+		  "var": "ext_wall_conductances",
+		  "bool_operator": "==",
+		  "expected_value": 0.183,
+		  "climate_index": 5
+		}
+	  ]
+	}
+  ]
+}
