+++ lib/openstudio-standards/standards/necb/necb_2011/qaqc/qaqc_data/hrv.json
@@ -1,0 +1,17 @@
+{
+  "tables": [
+    {
+      "name": "hrv_compliance",
+      "data_type": "table",
+      "refs": [
+        "NECB2011-5.2.10.1"
+      ],
+      "table": [
+	    {
+		  "var": "hrv_required?",
+		  "expected_value": "(0.00123*outdoor_air_L_per_s*(21-db990)) > 150"
+		}
+	  ]
+	}
+  ]
+}
