+++ lib/openstudio-standards/standards/necb/necb_2011/qaqc/qaqc_data/infiltration.json
@@ -1,0 +1,25 @@
+{
+  "tables": [
+    {
+      "name": "infiltration_compliance",
+      "data_type": "table",
+      "refs": [
+        "NECB2011-Section 8.4.3.6"
+      ],
+      "table": [
+	    {
+		  "var": "infiltration_method",
+		  "bool_operator": "==",
+		  "expected_value": "Flow/ExteriorArea",
+		  "tolerance": null
+		},
+		{
+		  "var": "infiltration_flow_per_m2",
+		  "bool_operator": "==",
+		  "expected_value": 0.00025,
+		  "tolerance": 5
+		}
+	  ]
+	}
+  ]
+}
