--- lib/openstudio-standards/standards/necb/necb_2011/qaqc_data/design_supply_temp.json
@@ -1,29 +1,0 @@
-{
-  "tables": [
-    {
-      "name": "design_supply_temp_compliance",
-      "data_type": "table",
-      "refs": [
-        "NECB2011-?"
-      ],
-	  "exclude_note": "The QAQC for a Thermal zone will be skipped if its name contains any of the substring(s) in the exclude_string",
-	  "exclude":{
-		"exclude_string": [
-		  "- undefined -"
-		]
-	  },
-      "table": [
-	    {
-		  "var": "heating_design_supply_air_temp",
-		  "bool_operator": "==",
-		  "expected_value": 43.0
-		},
-		{
-		  "var": "cooling_design_supply_temp",
-		  "bool_operator": "==",
-		  "expected_value": 13.0
-		}
-	  ]
-	}
-  ]
-}
