--- lib/openstudio-standards/standards/necb/necb_2011/qaqc_data/vav_fan_power.json
@@ -1,19 +1,0 @@
-{
-  "tables": [
-    {
-      "name": "vav_fan_power_compliance",
-      "data_type": "table",
-      "refs": [
-        "NECB2011-5.2.3.3"
-      ],
-	  "notes":"Do not change 'max_air_flow_rate_m3_per_s'",
-	  "formulas": {
-	    "NECB PSZ fan power (W)": "max_air_flow_rate_m3_per_s*1000*1.6",
-		"NECB VAV fan power (W)": "max_air_flow_rate_m3_per_s*1000*2.65"
-	  },
-      "table": [
-	    
-	  ]
-	}
-  ]
-}
